move TestListChannelsCmdF to channel_e2e_test.go
